+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
+- PBAP Client: allow to set start offset and max list count for pull phonebook operation
 
 ### Fixed
 - RFCOMM: shut down multiplexer after closing last channel instead of multiplexer idle timer